+++ segmodels_keras/backbones/_weights.py
@@ -5,7 +5,7 @@ Source: https://github.com/qubvel/classification_models
 
 from keras.utils import get_file
 
-from segmodels_keras._compat import load_weights
+from segmodels_keras.utils import load_weights
 
 __all__ = ["load_model_weights"]